[FIX] mail: fix disappearing moderated messages
task-2277748
task-2277811

X-original-commit: 5ef6aedb8271fa05f79fb6f93f679caa9cd418ff
