+++ addons/mail/views/assets.xml
@@ -244,7 +244,8 @@
                 <script type="text/javascript" src="/mail/static/src/components/chatter/chatter_tests.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/chatter_topbar/chatter_topbar_tests.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/composer/composer_tests.js"></script>
-                <script type="text/javascript" src="/mail/static/src/components/discuss/discuss_tests.js"></script>
+                <script type="text/javascript" src="/mail/static/src/components/discuss/tests/discuss_moderation_tests.js"></script>
+                <script type="text/javascript" src="/mail/static/src/components/discuss/tests/discuss_tests.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/file_uploader/file_uploader_tests.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/follow_button/follow_button_tests.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/follower/follower_tests.js"></script>
